requirements.txt
streamlit>=1.28.0
openai>=1.3.0
groq>=0.4.0
langchain>=0.1.0
email-validator>=2.1.0
phonenumbers>=8.13.0
python-dotenv>=1.0.0


